update FullyApplied condition
